query類を一元管理
