+++ app/javascript/pages/Tasks.tsx
@@ -2,17 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { useQuery, useMutation, gql } from "@apollo/client";
 import { useNavigate, useLocation, Link } from "react-router-dom";
 import { Alert } from "../components/Alert";
-
-const GET_TASKS = gql`
-  query GetTasks {
-    tasks {
-      id
-      name
-      description
-      createdAt
-    }
-  }
-`;
+import { GET_TASKS } from '../graphql/query'
 
 interface State {
   alert: string;
